change on everything

+++ pages/admin/categoryPage.js
@@ -31,14 +31,18 @@ const CategoryPage = () => {
 
     const handleButtonDelete = async (e, id) => {
         e.preventDefault();
-        try {
-            await fetch("/api/category/deletedata?id=" + id, {
-                method: "DELETE",
-                headers: { "Content-Type": "application/json" },
-            })
-            window.location.reload()
-        } catch (err) {
-            console.log(err)
+        if (confirm('Are you sure want to delete this ?')) {
+            try {
+                await fetch("/api/category/deletedata?id=" + id, {
+                    method: "DELETE",
+                    headers: { "Content-Type": "application/json" },
+                })
+                window.location.reload()
+            } catch (err) {
+                console.log(err)
+            }
+        }else{
+            console.log("not gonna happen!")
         }
     };
